turtle art + fixed warnings

+++ src/health_bars.cpp
@@ -16,8 +16,8 @@ void EntityHealthBar::render(const int scrollX, const int scrollY, SDL_Renderer*
     vec2<double> targetPos {static_cast<Entity*>(entity)->getCenter()};
     targetPos.y = static_cast<Entity*>(entity)->getPos().y - _dimensions.y - 3;
     targetPos.x -= 8;
-    SDL_Rect upperBar{(int)targetPos.x + 2 - scrollX, (int)targetPos.y + 1 - scrollY, _dimensions.x * _health / _maxHealth, 1};
-    SDL_Rect lowerBar{(int)targetPos.x + 2 - scrollX, (int)targetPos.y + 2 - scrollY, _dimensions.x * _health / _maxHealth, 1};
+    SDL_Rect upperBar{(int)targetPos.x + 2 - scrollX, (int)targetPos.y + 1 - scrollY, static_cast<int>(_dimensions.x * _health / _maxHealth), 1};
+    SDL_Rect lowerBar{(int)targetPos.x + 2 - scrollX, (int)targetPos.y + 2 - scrollY, static_cast<int>(_dimensions.x * _health / _maxHealth), 1};
     SDL_Color lightColor {Util::lerpColor(_redLight, _greenLight, _health / _maxHealth)};
     SDL_Color darkColor {Util::lerpColor(_redDark, _greenDark, _health / _maxHealth)};
     SDL_Rect Bar{(int)targetPos.x + 2 - scrollX, (int)targetPos.y + 1 - scrollY, _dimensions.x, 2};
